test(ui): stub useHighlights per case instead of restoring shared mock state
A factory returns a ready-to-run hook stub so later tests do not depend on a finally block resetting module mock state.
